NotFoundException uit domein gehaalt, tests toegevoegd en enkele bugs opgelost.

+++ code/src/main/java/nl/rubend/ipass/rest/PostService.java
@@ -17,6 +17,7 @@ public class PostService {
 	public Response newPost(@Context SecurityContext securityContext, @FormParam("pageId") String pageId, @FormParam("repliedTo") String repliedToId, @FormParam("text") String text, @FormParam("file") List<String> files) {
 		User user= (User) securityContext.getUserPrincipal();
 		Page page=Page.getPage(pageId);
+		if(page==null) return Response.status(Response.Status.NOT_FOUND).build();
 		if(!page.isLid(user)) return Response.status(Response.Status.UNAUTHORIZED).build();
 		Post post=new Post(user,page,repliedToId.equals("")?null:Post.getPost(repliedToId),text);
 		for(String fileID:files) {
@@ -27,9 +28,10 @@ public class PostService {
 	@GET
 	@Path("/{postId}")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-	public Response publicUserProfile(@PathParam("postId") String postId,@Context SecurityContext securityContext) {
+	public Response getPost(@PathParam("postId") String postId,@Context SecurityContext securityContext) {
 		User user= (User) securityContext.getUserPrincipal();
 		Post post=Post.getPost(postId);
+		if(post==null) throw new NotFoundException("pagina niet gevonden");
 		if(!post.getPage().isLid(user)) return Response.status(Response.Status.UNAUTHORIZED).build();
 		return Response.ok(post).build();
 	}
@@ -38,6 +40,7 @@ public class PostService {
 	public Response deletePost(@PathParam("postId") String postId,@Context SecurityContext securityContext) {
 		User user= (User) securityContext.getUserPrincipal();
 		Post post=Post.getPost(postId);
+		if(post==null) throw new NotFoundException("Pagina niet gevonden");
 		if(!(post.getUser().equals(user) || post.getPage().getOwner().equals(user))) return Response.status(Response.Status.UNAUTHORIZED).build();
 		post.delete();
 		return Response.ok().build();
@@ -48,6 +51,7 @@ public class PostService {
 	public Response vote(@PathParam("postId") String postId,@FormParam("vote") String vote,@Context SecurityContext securityContext) {
 		User user= (User) securityContext.getUserPrincipal();
 		Post post=Post.getPost(postId);
+		if(post==null) throw new NotFoundException("Pagina niet gevonden");
 		if(!post.getPage().isLid(user)) return Response.status(Response.Status.UNAUTHORIZED).build();
 		if(vote.equals("up")) new Vote(user,post,1);
 		else if(vote.equals("down")) new Vote(user,post,-1);
